Updated Amzn API & SearchItem Entity UPC types

+++ src/main/java/com/theironyard/entities/SearchItem.java
@@ -3,6 +3,8 @@ package com.theironyard.entities;
 import com.theironyard.api.AmazonUtil;
 import org.springframework.http.ResponseEntity;
 
+import java.math.BigInteger;
+
 
 public class SearchItem {
 
@@ -12,11 +14,11 @@ public class SearchItem {
 
     public String model;
 
-    public int upc;
+    public BigInteger upc;
 
     public String imageUrl;
 
-    public SearchItem(String title, String brand, String model, int upc, String imageUrl) {
+    public SearchItem(String title, String brand, String model, BigInteger upc, String imageUrl) {
         this.title = title;
         this.brand = brand;
         this.model = model;
